moved typingListener from LiveChat to Common
