+++ src/main/java/io/rocketchat/livechat/LiveChatAPI.java
@@ -2,6 +2,7 @@ package io.rocketchat.livechat;
 
 import io.rocketchat.common.data.rpc.RPC;
 import io.rocketchat.common.listener.ConnectListener;
+import io.rocketchat.common.listener.TypingListener;
 import io.rocketchat.common.network.Socket;
 import io.rocketchat.common.utils.Utils;
 import io.rocketchat.livechat.callback.*;
