+++ src/main/java/io/rocketchat/livechat/callback/SubscribeListener.java
@@ -6,7 +6,7 @@ import io.rocketchat.livechat.middleware.LiveChatStreamMiddleware;
 /**
  * Created by sachin on 12/6/17.
  */
-// TODO: 29/7/17 Future change to make it more generalized, removing type from the method and adding Boolean success and String subId to the code
+// TODO: 29/7/17 Future change to make it more generalized, removing type from the method and adding Boolean success and String subId to the code, moving to common like TypingListener
 public interface SubscribeListener extends Listener{
     /**
      *
