+++ src/main/java/io/rocketchat/livechat/middleware/LiveChatStreamMiddleware.java
@@ -3,7 +3,7 @@ package io.rocketchat.livechat.middleware;
 import io.rocketchat.livechat.callback.AgentListener;
 import io.rocketchat.livechat.callback.MessageListener;
 import io.rocketchat.livechat.callback.SubscribeListener;
-import io.rocketchat.livechat.callback.TypingListener;
+import io.rocketchat.common.listener.TypingListener;
 import io.rocketchat.livechat.model.AgentObject;
 import io.rocketchat.livechat.model.LiveChatMessage;
 import org.json.JSONArray;
